refactor: 도메인별 에러코드를 도메인 하위 구조로 변경 - CategoryErrorCode - MenuErrorCode - ServiceErrorCode
