+++ src/main/java/com/team6/backend/menu/application/service/MenuService.java
@@ -2,7 +2,7 @@ package com.team6.backend.menu.application.service;
 
 import com.team6.backend.global.infrastructure.config.security.util.SecurityUtils;
 import com.team6.backend.global.infrastructure.exception.ApplicationException;
-import com.team6.backend.global.infrastructure.exception.MenuErrorCode;
+import com.team6.backend.menu.domain.exception.MenuErrorCode;
 import com.team6.backend.global.infrastructure.util.AuthValidator;
 import com.team6.backend.menu.domain.entity.Menu;
 import com.team6.backend.menu.domain.repository.MenuRepository;
