+++ src/main/java/com/team6/backend/store/application/service/StoreService.java
@@ -2,7 +2,7 @@ package com.team6.backend.store.application.service;
 
 import com.team6.backend.global.infrastructure.config.security.util.SecurityUtils;
 import com.team6.backend.global.infrastructure.exception.ApplicationException;
-import com.team6.backend.global.infrastructure.exception.StoreErrorCode;
+import com.team6.backend.store.domain.exception.StoreErrorCode;
 import com.team6.backend.global.infrastructure.util.AuthValidator;
 import com.team6.backend.store.domain.entity.Store;
 import com.team6.backend.store.domain.repository.StoreRepository;
